Move MetadataResolver to presto-spi
We need to use metadataResolver in the pluggable analyzer interfaces,
which is present in the presto-spi, hence we are moving
metadataResolver and ViewDefinition to presto-spi.
